Submitting minor changes

+++ 1. C# Basics/7. Programming Basics C# Book/LoopsAndConditions/ChangeTiles/Program.cs
@@ -13,19 +13,18 @@ namespace ChangeTiles
             decimal tileHeight = decimal.Parse(Console.ReadLine());
             decimal tileCost = decimal.Parse(Console.ReadLine());
             decimal workerWage = decimal.Parse(Console.ReadLine());
-            decimal floorArea = (decimal)(floorSideA * floorSideB);
-            decimal tileArea = (decimal)(tileBase * tileHeight / 2);
+            decimal floorArea = (floorSideA * floorSideB);
+            decimal tileArea = (tileBase * tileHeight / 2);
             decimal tileCount = Math.Ceiling(floorArea / tileArea) + 5;
-            decimal totalCost = (decimal)(tileCount * tileCost + workerWage);
-            if (savedMoney-totalCost>=0)
+            decimal totalCost = (tileCount * tileCost + workerWage);
+            if (savedMoney - totalCost >= 0)
             {
-                Console.WriteLine("{0:F2} lv left.",savedMoney-totalCost);
+                Console.WriteLine("{0:F2} lv left.", savedMoney - totalCost);
             }
             else
             {
-                Console.WriteLine("You'll need {0:F2} lv more.", totalCost-savedMoney);
+                Console.WriteLine("You'll need {0:F2} lv more.", totalCost - savedMoney);
             }
-
         }
     }
 }
